% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{bar_plot}
\alias{bar_plot}
\title{Plot the Number of Species by Invasion Status}
\usage{
bar_plot()
}
\value{
A \code{ggplot} object representing the number of species in each invasion
status category.
}
\description{
Retrieves species data from the \code{sp_categorization} table in the ILORA database
and creates a bar plot showing the number of species in each invasion status category.
Categories include Invasive (In), Naturalized (Nt), Casual Aliens (CA),
Cryptogenic (CG), and Native (N).
}
\examples{
\donttest{
# Generate a bar plot of species counts for each invasion status
bar_plot()
}

}
