% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_species_names}
\alias{get_species_names}
\title{Retrieve Species Names from ILORA Database}
\usage{
get_species_names()
}
\value{
A character vector containing all species names in the ILORA database.
If no species are found, returns an empty character vector.
}
\description{
This function connects to the ILORA database and retrieves all species names
available in the \code{species_name} table. It returns them as a character vector
in alphabetical order.
}
\examples{
\donttest{
species_names <- get_species_names()
head(species_names)
}

}
