% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{ILORA}
\alias{ILORA}
\title{ILORA: A package for retrieving species data from ILORA database}
\description{
Provides tools to access, query, and visualize species occurrence data from the Indian Alien Flora Information (ILORA) database.
The package enables users to retrieve species names, explore available variables, and extract species-level data based on user-defined criteria.
It also supports exploratory data analysis and visualization to facilitate ecological and biogeographical research.
Further details about the database are available at \url{https://ilora2020.wixsite.com/ilora2020}.
}
