% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{ilora_variable_guide}
\alias{ilora_variable_guide}
\title{Open ILORA Variable Description PDF}
\usage{
ilora_variable_guide()
}
\value{
No return value. This function is called for its side effect of
opening the PDF file.
}
\description{
Opens the ILORA variable description PDF file (\code{ILORA_HowToRead.pdf})
in the system's default PDF viewer. This guide helps users understand
the variables and structure of the ILORA dataset.
}
\details{
This function uses \code{utils::browseURL()} to open the PDF file.
Depending on the user's system configuration, the file may open in
a web browser, external PDF viewer, or within RStudio's viewer pane.
}
\examples{
\donttest{
ilora_variable_guide()
}

}
