% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{introduction_pathways_plot}
\alias{introduction_pathways_plot}
\title{Species Count Plot by Introduction Pathways}
\usage{
introduction_pathways_plot()
}
\value{
A \code{plotly} object representing a 3D pie chart of species counts by introduction pathway.
The summary table of top pathways is printed as a message.
}
\description{
Retrieves species data from the \code{introduction} table in the ILORA database,
counts distinct species for each reported introduction pathway, and generates a 3D pie chart
using Plotly. The summary table of counts is also printed.
}
\examples{
\donttest{
# Generate pie chart for species introduction pathways
plot <- introduction_pathways_plot()
plot  # Displays the Plotly pie chart
}

}
