% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{plot_first_record_year_histogram}
\alias{plot_first_record_year_histogram}
\title{Histogram of Species First Record Year}
\usage{
plot_first_record_year_histogram(start_year, end_year, binwidth = NULL)
}
\arguments{
\item{start_year}{Numeric. Start year of the interval.}

\item{end_year}{Numeric. End year of the interval.}

\item{binwidth}{Numeric, optional. Width of histogram bins. If \code{NULL} (default), it
is automatically determined based on the year range.}
}
\value{
A \code{ggplot} object representing the histogram of species first record years.
}
\description{
Retrieves first record years of species from the \code{introduction} table
in the ILORA database within a specified year interval and plots a histogram using ggplot2.
}
\examples{
\donttest{
# Plot histogram for species first records between 1990 and 2000
plot_first_record_year_histogram(start_year = 1990, end_year = 2000)
}

}
