% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\docType{package}
\name{SpeciesDataPackage}
\alias{-package}
\alias{SpeciesDataPackage}
\title{SpeciesDataPackage: A package for retrieving species data from ILORA database}
\description{
This package provides functions to interact with a ILORA database to retrieve
species data based on species name and specified variables.
}
