% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{visualize_native_range}
\alias{visualize_native_range}
\title{Visualize Native Range of Species on a Global Map}
\usage{
visualize_native_range(species_name)
}
\arguments{
\item{species_name}{Character vector of species names to plot. Use \code{"ALL"} to plot all species in the database.}
}
\value{
A ggplot2 object showing the native range of the specified species.
}
\description{
This function retrieves the native range of one or more species from the ILORA database
and plots their distribution on a global map using TDWG Level 2 regions.
}
\details{
The function joins the native range data from the ILORA database with a TDWG Level 2 shapefile
to display species’ native regions. Each species is color-coded, and overlapping ranges are handled
to avoid duplicate polygons.
}
\examples{
\donttest{
# Visualize native range of specific species
visualize_native_range(c("Cyanus segetum Hill", "Avena sativa L."))
}

}
