% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{visualize_naturalized_range}
\alias{visualize_naturalized_range}
\title{Visualize Naturalized Range of Species on a Global Map}
\usage{
visualize_naturalized_range(species_name)
}
\arguments{
\item{species_name}{Character vector of species names to plot. Use \code{"ALL"} to plot all species in the database.}
}
\value{
A ggplot2 object showing the naturalized range of the specified species.
}
\description{
This function retrieves the naturalized range of one or more species from the ILORA database
and plots their distribution on a global map using TDWG Level 2 regions.
}
\details{
The function joins the naturalized range data from the ILORA database with a TDWG Level 2 shapefile
to display species’ naturalized regions. Each species is color-coded.
}
\examples{
\donttest{
# Visualize naturalized range of specific species
visualize_naturalized_range(c("Cyanus segetum Hill", "Avena sativa L."))
}

}
