% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Retrieve Data for a Given Species from the ILORA Database}
\usage{
get_data(species_name, variables)
}
\arguments{
\item{species_name}{A character string representing the species name to search for in the database tables.}

\item{variables}{A character vector specifying the column names to retrieve from the tables.}
}
\value{
A data frame containing the retrieved data, or NULL if no data is found.
}
\description{
This function searches for a specified species within all tables of the ILORA database schema and retrieves the desired variables.
The search is performed across all relevant tables to ensure comprehensive data retrieval for the specified species.
}
\details{
\subsection{To fetch Species names available in Database}{

All species names available in database can be retrieved using the function:
\code{get_species_names()}
}

\subsection{To fetch Variable names available in database}{

All variable names available in database can be retrieved using the function:
\code{get_variable_names()}

The following are examples of species names and variable names that can be used:
\itemize{
\item Species names: 'Rotala densiflora (Roth) Koehne','Rotala indica (Willd.) Koehne','Rubus buergeri Miq.',
\item Variable names: "orders", "0700_Fuels2", "genus", "species"
}
}
}
