% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_species_names}
\alias{get_species_names}
\title{Retrieve Species Names}
\usage{
get_species_names()
}
\value{
A character vector containing species names.
}
\description{
This function returns a vector of species names available in ILROA database.
}
