% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_variable_names}
\alias{get_variable_names}
\title{Retrieve Variable Names}
\usage{
get_variable_names()
}
\value{
A character vector containing variable names.
}
\description{
This function returns a vector of variable names available in ILROA database.
}
